Improve filtering of actuator auto-configurations
This commit splits auto-configurations that require different
environments so that they can be filtered early.

Closes gh-12260
