+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/main/java/org/springframework/boot/actuate/autoconfigure/redis/RedisHealthIndicatorAutoConfiguration.java
@@ -16,30 +16,54 @@
 
 package org.springframework.boot.actuate.autoconfigure.redis;
 
+import java.util.Map;
+
+import org.springframework.boot.actuate.autoconfigure.health.CompositeHealthIndicatorConfiguration;
 import org.springframework.boot.actuate.autoconfigure.health.ConditionalOnEnabledHealthIndicator;
 import org.springframework.boot.actuate.autoconfigure.health.HealthIndicatorAutoConfiguration;
+import org.springframework.boot.actuate.health.HealthIndicator;
 import org.springframework.boot.actuate.redis.RedisHealthIndicator;
-import org.springframework.boot.actuate.redis.RedisReactiveHealthIndicator;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
 
 /**
- * {@link EnableAutoConfiguration Auto-configuration} for {@link RedisHealthIndicator} and
- * {@link RedisReactiveHealthIndicator}.
+ * {@link EnableAutoConfiguration Auto-configuration} for {@link RedisHealthIndicator}.
  *
- * @author Phillip Webb
- * @since 2.0.0
+ * @author Christian Dupuis
+ * @author Richard Santana
+ * @author Stephane Nicoll
+ * @author Mark Paluch
+ * @since 2.1.0
  */
 @Configuration
+@ConditionalOnClass(RedisConnectionFactory.class)
+@ConditionalOnBean(RedisConnectionFactory.class)
 @ConditionalOnEnabledHealthIndicator("redis")
 @AutoConfigureBefore(HealthIndicatorAutoConfiguration.class)
-@AutoConfigureAfter(RedisAutoConfiguration.class)
-@Import({ RedisReactiveHealthIndicatorConfiguration.class,
-		RedisHealthIndicatorConfiguration.class })
-public class RedisHealthIndicatorAutoConfiguration {
+@AutoConfigureAfter({ RedisAutoConfiguration.class,
+		RedisReactiveHealthIndicatorAutoConfiguration.class })
+public class RedisHealthIndicatorAutoConfiguration extends
+		CompositeHealthIndicatorConfiguration<RedisHealthIndicator, RedisConnectionFactory> {
+
+	private final Map<String, RedisConnectionFactory> redisConnectionFactories;
+
+	public RedisHealthIndicatorAutoConfiguration(
+			Map<String, RedisConnectionFactory> redisConnectionFactories) {
+		this.redisConnectionFactories = redisConnectionFactories;
+	}
+
+	@Bean
+	@ConditionalOnMissingBean(name = "redisHealthIndicator")
+	public HealthIndicator redisHealthIndicator() {
+		return createHealthIndicator(this.redisConnectionFactories);
+	}
 
 }
