--- spring-boot-project/spring-boot-actuator-autoconfigure/src/main/java/org/springframework/boot/actuate/autoconfigure/redis/RedisHealthIndicatorConfiguration.java
@@ -1,58 +1,0 @@
-/*
- * Copyright 2012-2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.boot.actuate.autoconfigure.redis;
-
-import java.util.Map;
-
-import org.springframework.boot.actuate.autoconfigure.health.CompositeHealthIndicatorConfiguration;
-import org.springframework.boot.actuate.health.HealthIndicator;
-import org.springframework.boot.actuate.redis.RedisHealthIndicator;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-
-/**
- * Configuration for {@link RedisHealthIndicator}.
- *
- * @author Christian Dupuis
- * @author Richard Santana
- * @author Stephane Nicoll
- * @author Mark Paluch
- */
-@Configuration
-@ConditionalOnClass(RedisConnectionFactory.class)
-@ConditionalOnBean(RedisConnectionFactory.class)
-class RedisHealthIndicatorConfiguration extends
-		CompositeHealthIndicatorConfiguration<RedisHealthIndicator, RedisConnectionFactory> {
-
-	private final Map<String, RedisConnectionFactory> redisConnectionFactories;
-
-	RedisHealthIndicatorConfiguration(
-			Map<String, RedisConnectionFactory> redisConnectionFactories) {
-		this.redisConnectionFactories = redisConnectionFactories;
-	}
-
-	@Bean
-	@ConditionalOnMissingBean(name = "redisHealthIndicator")
-	public HealthIndicator redisHealthIndicator() {
-		return createHealthIndicator(this.redisConnectionFactories);
-	}
-
-}
