+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/test/java/org/springframework/boot/actuate/autoconfigure/couchbase/CouchbaseReactiveHealthIndicatorAutoConfigurationTests.java
@@ -15,6 +15,8 @@
  */
 package org.springframework.boot.actuate.autoconfigure.couchbase;
 
+import java.time.Duration;
+
 import org.junit.Test;
 
 import org.springframework.boot.actuate.autoconfigure.health.HealthIndicatorAutoConfiguration;
@@ -25,30 +27,29 @@ import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.couchbase.core.CouchbaseOperations;
-import org.springframework.test.util.ReflectionTestUtils;
+import org.springframework.data.couchbase.core.RxJavaCouchbaseOperations;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
 /**
- * Tests for {@link CouchbaseHealthIndicatorConfiguration}.
+ * Tests for {@link CouchbaseReactiveHealthIndicatorAutoConfiguration}.
  *
- * @author Phillip Webb
- * @author Stephane Nicoll
+ * @author Mikalai Lushchytski
  */
-public class CouchbaseHealthIndicatorConfigurationTests {
+public class CouchbaseReactiveHealthIndicatorAutoConfigurationTests {
 
 	private ApplicationContextRunner contextRunner = new ApplicationContextRunner()
-			.withUserConfiguration(CouchbaseMockConfiguration.class).withConfiguration(
-					AutoConfigurations.of(CouchbaseHealthIndicatorAutoConfiguration.class,
-							HealthIndicatorAutoConfiguration.class));
+			.withUserConfiguration(CouchbaseMockConfiguration.class)
+			.withConfiguration(AutoConfigurations.of(
+					CouchbaseReactiveHealthIndicatorAutoConfiguration.class,
+					HealthIndicatorAutoConfiguration.class));
 
 	@Test
 	public void runShouldCreateIndicator() {
 		this.contextRunner.run((context) -> assertThat(context)
-				.hasSingleBean(CouchbaseHealthIndicator.class)
-				.doesNotHaveBean(CouchbaseReactiveHealthIndicator.class)
+				.hasSingleBean(CouchbaseReactiveHealthIndicator.class)
+				.doesNotHaveBean(CouchbaseHealthIndicator.class)
 				.doesNotHaveBean(ApplicationHealthIndicator.class));
 	}
 
@@ -56,10 +57,11 @@ public class CouchbaseHealthIndicatorConfigurationTests {
 	public void runWithCustomTimeoutShouldCreateIndicator() {
 		this.contextRunner.withPropertyValues("management.health.couchbase.timeout=2s")
 				.run((context) -> {
-					assertThat(context).hasSingleBean(CouchbaseHealthIndicator.class);
-					assertThat(ReflectionTestUtils.getField(
-							context.getBean(CouchbaseHealthIndicator.class), "timeout"))
-									.isEqualTo(2000L);
+					assertThat(context)
+							.hasSingleBean(CouchbaseReactiveHealthIndicator.class);
+					assertThat(context.getBean(CouchbaseReactiveHealthIndicator.class))
+							.hasFieldOrPropertyWithValue("timeout",
+									Duration.ofSeconds(2));
 				});
 	}
 
@@ -67,7 +69,7 @@ public class CouchbaseHealthIndicatorConfigurationTests {
 	public void runWhenDisabledShouldNotCreateIndicator() {
 		this.contextRunner.withPropertyValues("management.health.couchbase.enabled:false")
 				.run((context) -> assertThat(context)
-						.doesNotHaveBean(CouchbaseHealthIndicator.class)
+						.doesNotHaveBean(CouchbaseReactiveHealthIndicator.class)
 						.hasSingleBean(ApplicationHealthIndicator.class));
 	}
 
@@ -75,8 +77,8 @@ public class CouchbaseHealthIndicatorConfigurationTests {
 	protected static class CouchbaseMockConfiguration {
 
 		@Bean
-		public CouchbaseOperations couchbaseOperations() {
-			return mock(CouchbaseOperations.class);
+		public RxJavaCouchbaseOperations couchbaseOperations() {
+			return mock(RxJavaCouchbaseOperations.class);
 		}
 
 	}
